+++ doc/CHANGELOG.rst
@@ -5,6 +5,9 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
 
 - **Next version - unreleased**
 
+  - Corrected an issue with creation of Exceptions which lack a
+    default constructor.
+
 - **0.7.2 - 2-28-2019**
   - C++ and Java exceptions hold the traceback as a Python exception
     cause.  It is no longer necessary to call stacktrace() to retrieve